backend/pyproject.toml
[project]
name = "secure-backend"
version = "0.1.0"
description = "Secure corporate chat backend (E2EE routing only)"
requires-python = ">=3.11"

dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "sqlalchemy==2.0.30",
  "psycopg[binary]==3.1.19",
  "pydantic==2.7.4",
  "python-multipart==0.0.9",

  # password hashing (fix passlib+bcrypt compatibility)
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.0.1",

  "python-jose==3.3.0",
  "aiofiles==23.2.1",
]

[tool.uvicorn]
factory = false